LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/fetch/sec-metadata/embed.tentative.https.sub-expected.txt



FAIL Same-Origin embed assert_not_equals: Empty Sec-Metadata header. got disallowed value ""
FAIL Same-Site embed assert_equals: destination expected (string) "embed" but got (undefined) undefined
FAIL Cross-Site embed assert_equals: destination expected (string) "embed" but got (undefined) undefined
